Commit after adding ccsds secondary header and implementing can_run()

+++ src/PiLOT_sb_MSS_CM3_app/pilot.c
@@ -444,7 +444,7 @@ uint8_t sd_hk_test(sd_test *sd,uint8_t *data,uint32_t addr,uint8_t *sd_state) {
 		if(sd->sd_result == 0) {
 			MSS_GPIO_set_output(SD_CARD_GPIO,0);
 			*sd_state = 0;
-			uint32_t ph,pl;
+			uint64_t ph,pl;
 			time_to_count(60000,&ph,&pl);
 			TMR_init(&sd_timer,SD_TIMER_BASE_ADDR,TMR_ONE_SHOT_MODE,PRESCALER_DIV_2,pl/2);
 			TMR_enable_int(&sd_timer);
@@ -456,7 +456,7 @@ uint8_t sd_hk_test(sd_test *sd,uint8_t *data,uint32_t addr,uint8_t *sd_state) {
 	return 0;
 }
 
-void time_to_count(uint32_t ms,uint32_t *upper_count,uint32_t *lower_count) {
+void time_to_count(uint32_t ms,uint64_t *upper_count,uint64_t *lower_count) {
     *lower_count = (ms%FULL_SCALE_TIME_MS) * TIMER_COUNT_PER_MS;
     *upper_count = (ms/FULL_SCALE_TIME_MS);
 }
